Add inline code event
