+++ platforms/web/src/src/lib/types.ts
@@ -1,0 +1,2 @@
+export type BlockType = 'formatInlineCode';
+export type WysiwygInputEvent = InputEvent & { inputType: InputEvent['inputType'] | BlockType };
